refactor: swagger 문서화 수정 및 불필요한 코드 제거

+++ src/main/java/yerong/wedle/timetable/personalTimetable/service/PersonalTimetableService.java
@@ -63,6 +63,9 @@ public class PersonalTimetableService {
         PersonalTimetable personalTimetable = personalTimetableRepository.findByMember(member)
                 .orElseThrow(
                         PersonalTimetableNotFoundException::new);
+        if (!personalTimetable.getMember().getSocialId().equals(socialId)) {
+            throw new UnauthorizedAccessException();
+        }
         personalScheduleRepository.deleteInBatch(personalTimetable.getSchedules());
     }
